requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.7
django-appconf==1.0.5
django-bootstrap-v5==1.0.5
django-cors-headers==3.10.0
django-extensions==3.1.3
django-imagekit==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pilkit==2.0
Pillow==8.3.2
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
